+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated types
 - Updated setup to use isort
 - Updated local dev setup
+- Updated platforms to have available option and unique IDs
 
 ## 0.3.0